+++ R/api_endpoints.R
@@ -24,7 +24,7 @@
 #'   which a specific set of indicators is present, add
 #'   `selectSurveys="byIndicator"` in conjunction with Indicator IDs, Country
 #'   Code, and/or Survey Type instead
-#'   of using `selectSurveys="latest"`.
+#'   of using `selectSurveys="latest"`
 #' @param surveyYear Specify a comma separated list of survey years to
 #'   filter by.
 #' @param surveyYearStart Specify a range of Survey Years to filter Data on.