+++ R/read_datasets.R
@@ -173,8 +173,7 @@ factor_format <- function(res, reformat=FALSE, all_lower=TRUE) {
       pos <- match(names(lab[i]), names(res))
       lab_match <- which(!is.na(match(res[[pos]], lab[[i]])))
       if (length(lab_match) > 0) {
-        res[[pos]][lab_match] <- names(lab[[i]])[match(res[[pos]][lab_match],
-                                                       lab[[i]])]
+        res[[pos]] <- haven::as_factor(res[[pos]])
       }
     }
 
@@ -187,8 +186,6 @@ factor_format <- function(res, reformat=FALSE, all_lower=TRUE) {
                                            "label",
                                            description_table$description)
   }
-
-  class(res) <- c(class(res), "dhs_dataset")
 
   return(list("dataset" = res, "variable_names" = description_table))
 }