+++ R/read_dhs_dta.R
@@ -12,12 +12,8 @@
 #'
 #' @examples
 #' mrdt_zip <- tempfile()
-#' download.file(
-#' paste0(
-#' "https://dhsprogram.com/customcf/legacy/data/sample_download_dataset.cfm",
-#' "?Filename=ZZMR61DT.ZIP&Tp=1&Ctry_Code=zz&survey_id=0&doctype=dhs"),
-#' mrdt_zip, mode="wb"
-#' )
+#' download.file("https://dhsprogram.com/data/model_data/dhs/zzmr61fl.zip",
+#'               mrdt_zip, mode="wb")
 #'
 #' map <- rdhs::read_zipdata(mrdt_zip, "\\.MAP", readLines)
 #' dct <- rdhs:::parse_map(map)
@@ -216,13 +212,9 @@ parse_map <- function(map, all_lower=TRUE) {
 #'
 #' @examples
 #' mrdt_zip <- tempfile()
-#' download.file(
-#' paste0(
-#' "https://dhsprogram.com/customcf/legacy/data/sample_download_dataset.cfm?",
-#' "Filename=ZZMR61DT.ZIP&Tp=1&Ctry_Code=zz&survey_id=0&doctype=dhs"),
-#' mrdt_zip, mode="wb"
-#' )
-#'
+#' download.file("https://dhsprogram.com/data/model_data/dhs/zzmr61dt.zip",
+#'               mrdt_zip, mode="wb")
+#' 
 #' mr <- rdhs::read_dhs_dta(mrdt_zip,mode="map")
 #' attr(mr$mv213, "label")
 #' class(mr$mv213)