+++ tests/testthat/test_downloads.R
@@ -27,7 +27,7 @@ test_that("available surveys and download work", {
   )
 
   # check handle for dataset you don't have permission for
-  downloads <- cli$get_datasets("CDGC61FL.ZIP")
+  expect_message(downloads <- cli$get_datasets("CDGC61FL.ZIP"), "not available")
 
   # check dta foreign only
   downloads <- cli$get_datasets(
@@ -134,7 +134,7 @@ test_that("ETAR71FL.ZIP test", {
 
   dat <- cli$get_datasets("ETAR71FL.ZIP")
   r <- readRDS(dat[[1]])
-  expect_identical(class(r), c("data.frame", "dhs_dataset"))
+  expect_identical(class(r), "data.frame")
 })
 
 test_that("ugir41fl.zip test", {