+++ tests/testthat/test_read_dhs_flat.R
@@ -3,12 +3,8 @@ context("Parse datasets")
 test_that("datasets parse", {
   mrfl_zip <- tempfile()
   on.exit(unlink(mrfl_zip))
-  download.file(paste0(
-    "https://dhsprogram.com/customcf/legacy/data/sample_download_dataset.cfm?",
-    "Filename=ZZMR61FL.ZIP&Tp=1&Ctry_Code=zz&survey_id=0&doctype=dhs"
-  ),
-  destfile = mrfl_zip, mode = "wb"
-  )
+  download.file("https://dhsprogram.com/data/model_data/dhs/zzmr61fl.zip",
+                destfile = mrfl_zip, mode = "wb")
 
   mr_def <- read_dhs_flat(mrfl_zip)
   mr_dcf <- read_dhs_flat(mrfl_zip, meta_source = "dcf")
@@ -49,15 +45,13 @@ test_that("datasets parse", {
 test_that("data dictionaries FWF lengths match file width", {
   arfl_zip <- tempfile()
   on.exit(unlink(arfl_zip))
-  download.file(paste0(
-    "https://dhsprogram.com/customcf/legacy/data/sample_download_dataset.cfm?",
-    "Filename=ZZAR61FL.ZIP&Tp=4&Ctry_Code=zz&survey_id=0&doctype=hiv"
-  ), arfl_zip, mode = "wb")
+  download.file("https://dhsprogram.com/data/model_data/hiv/zzar61fl.zip",
+                arfl_zip, mode = "wb")
 
-  dcf <- rdhs::read_zipdata(arfl_zip, "\\.DCF", readLines)
-  sps <- rdhs::read_zipdata(arfl_zip, "\\.SPS", readLines)
-  do <- rdhs::read_zipdata(arfl_zip, "\\.DO", readLines)
-  dct <- rdhs::read_zipdata(arfl_zip, "\\.DCT", readLines)
+  dcf <- rdhs::read_zipdata(arfl_zip, "\\.DCF", brio::read_lines)
+  sps <- rdhs::read_zipdata(arfl_zip, "\\.SPS", brio::read_lines)
+  do <- rdhs::read_zipdata(arfl_zip, "\\.DO", brio::read_lines)
+  dct <- rdhs::read_zipdata(arfl_zip, "\\.DCT", brio::read_lines)
 
   dat <- rdhs::read_zipdata(arfl_zip, "\\.DAT$", iotools::input.file)
 
@@ -66,7 +60,7 @@ test_that("data dictionaries FWF lengths match file width", {
   expect_equal(sum(parse_do(do, dct)$len), nchar(dat[1]))
 
   # check for incorrect pattern
-  expect_warning(rdhs::read_zipdata(arfl_zip, "\\.notachance", readLines))
+  expect_warning(rdhs::read_zipdata(arfl_zip, "\\.notachance", brio::read_lines))
 })