+++ tests/testthat/test_spatial_boundaries.R
@@ -15,7 +15,8 @@ test_that("Spatial Boundaries Download", {
    # using the surveyNum
    res <- download_boundaries(surveyNum = 471, countryId = "AF")
    expect_true(res$sdr_subnational_boundaries$ISO[1] == "AF")
-
+   expect_is(res$sdr_subnational_boundaries, "sf")
+  
    # using the surveyId and no countryID
    res <- download_boundaries(surveyId = "AF2010OTH")
    expect_true(length(res) == 2)
@@ -24,6 +25,19 @@ test_that("Spatial Boundaries Download", {
    # using sf
    res <- download_boundaries(surveyNum = 471, countryId = "AF", method = "sf")
    expect_true(res$sdr_subnational_boundaries$ISO[1] == "AF")
+   expect_is(res$sdr_subnational_boundaries, "sf")
+
+   # using rgdal
+   res <- download_boundaries(surveyNum = 471, countryId = "AF", method = "rgdal")
+   expect_true(res$sdr_subnational_boundaries$ISO[1] == "AF")
+   expect_is(res$sdr_subnational_boundaries, "SpatialPolygonsDataFrame")
+
+   # unsupported method
+    expect_message(res <- download_boundaries(surveyNum = 471, countryId = "AF",
+                                              method = "jibberish"),
+                   "Provided method not found.")
+    expect_is(res, "character")
+    expect_true(any(grepl("\\.shp$", res)))
 
 })